Make it possible to replace multiple lines
Fixes #2

+++ find-and-replace-maven-plugin/src/main/java/software/xdev/far/filecontents/FileContentsProcessor.java
@@ -9,6 +9,7 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.io.UncheckedIOException;
+import java.nio.file.Files;
 import java.util.regex.Matcher;
 
 import software.xdev.far.BaseProcessor;
@@ -26,51 +27,73 @@ public class FileContentsProcessor extends BaseProcessor<FileContentsExecData>
 	{
 		try
 		{
-			final File tempFile = File.createTempFile("tmp", "tmp", file.getParentFile());
+			File tempFile = null;
 			
-			try(final FileInputStream fis = new FileInputStream(file);
-				final InputStreamReader isr = new InputStreamReader(fis, this.execData.getCharset());
-				final BufferedReader fileReader = new BufferedReader(isr))
+			if(this.execData.isReplaceLineBased())
 			{
-				try(final FileOutputStream fos = new FileOutputStream(tempFile);
-					final OutputStreamWriter osr = new OutputStreamWriter(fos, this.execData.getCharset());
-					final BufferedWriter fileWriter = new BufferedWriter(osr))
+				tempFile = this.createTempFile(file);
+				try(final FileInputStream fis = new FileInputStream(file);
+					final InputStreamReader isr = new InputStreamReader(fis, this.execData.getCharset());
+					final BufferedReader fileReader = new BufferedReader(isr))
 				{
-					boolean alreadyReplaced = false;
-					
-					for(String line = fileReader.readLine(); line != null; line = fileReader.readLine())
+					try(final FileOutputStream fos = new FileOutputStream(tempFile);
+						final OutputStreamWriter osr = new OutputStreamWriter(fos, this.execData.getCharset());
+						final BufferedWriter fileWriter = new BufferedWriter(osr))
 					{
-						final Matcher matcher = this.execData.getFindRegex().matcher(line);
-						if(matcher.find())
+						boolean alreadyReplaced = false;
+						for(String line = fileReader.readLine(); line != null; line = fileReader.readLine())
 						{
-							if(this.execData.isReplaceAll())
+							final Matcher matcher = this.execData.getFindRegex().matcher(line);
+							if(matcher.find())
 							{
-								line = matcher.replaceAll(this.execData.getReplaceValue());
+								if(this.execData.isReplaceAll())
+								{
+									line = matcher.replaceAll(this.execData.getReplaceValue());
+								}
+								else if(!alreadyReplaced)
+								{
+									line = matcher.replaceFirst(this.execData.getReplaceValue());
+									alreadyReplaced = true;
+								}
 							}
-							else if(!alreadyReplaced)
-							{
-								line = matcher.replaceFirst(this.execData.getReplaceValue());
-								alreadyReplaced = true;
-							}
+							fileWriter.write(line + System.lineSeparator());
 						}
-						fileWriter.write(line + System.lineSeparator());
 					}
 				}
 			}
-			
-			if(!file.delete())
+			else
 			{
-				throw new IOException("Failed to delete file at: " + file.getPath());
+				final String contents = Files.readString(file.toPath(), this.execData.getCharset());
+				final Matcher matcher = this.execData.getFindRegex().matcher(contents);
+				if(matcher.find())
+				{
+					tempFile = this.createTempFile(file);
+					
+					Files.writeString(tempFile.toPath(), this.execData.isReplaceAll()
+						? matcher.replaceAll(this.execData.getReplaceValue())
+						: matcher.replaceFirst(this.execData.getReplaceValue()));
+				}
 			}
 			
-			if(!tempFile.renameTo(file))
+			if(tempFile != null)
 			{
-				throw new IOException("Failed to rename temp file at: " + tempFile.getPath() + " to " + file.getPath());
+				Files.delete(file.toPath());
+				
+				if(!tempFile.renameTo(file))
+				{
+					throw new IOException(
+						"Failed to rename temp file at: " + tempFile.getPath() + " to " + file.getPath());
+				}
 			}
 		}
 		catch(final IOException e)
 		{
 			throw new UncheckedIOException(e);
 		}
+	}
+	
+	protected File createTempFile(final File original) throws IOException
+	{
+		return File.createTempFile("tmp", "tmp", original.getParentFile());
 	}
 }
